+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * new: All Assembly\*Adapter add getItemFactoryByData(Any) and getItemFactoryByItemFactoryClass(KClass) methods
 * new: ItemDataStorage.onDataListChanged function adds `oldList: List<DATA>` and `newList: List<DATA>` parameters
 * improve: Improve the implementation of AssemblySingleData\*Adapter
+* new: AssemblySingleData\*Adapter adds itemFactory property
 
 # v4.0.0-rc04
 * restore: Restore ArrayFragmentStateAdapter, ArrayPagerAdapter, ArrayFragmentStatePagerAdapter